fix: resolve pickle CRLF, medcoupling int32, and numpy compat issues on Windows
- variable.py: strip CRLF from pickle files — Windows text-mode transfers
  corrupt the pickle STRING opcode which expects LF-only line endings
- field_converter.py: cast np.where() int64 arrays to int32 for medcoupling
  DataArrayInt which requires INT32 on all platforms
- post_bordet_ops.py: use .tolist() instead of list comprehension for
  converting numpy arrays to native Python types for CREA_TABLE
- post_endo_fiss_utils.py: use .item() instead of float() on 1-element
  numpy arrays — float() rejects non-0-dim arrays in numpy >= 2.0
- shll102a.comm: use NP.float64(x).item() for robust scalar extraction

+++ code_aster/MedUtils/MedConverter/field_converter.py
@@ -112,7 +112,9 @@ def toMCFieldAndProfileNode(asfield, medmesh, prefix=""):
     medc_node_field.setNature(medc.IntensiveMaximum)
 
     # Med profile
-    field_profile = medc.DataArrayInt(restricted_nodes)
+    # medcoupling DataArrayInt expects INT32 numpy arrays, but np.where() returns
+    # int64 on 64-bit platforms. Cast explicitly to int32 for compatibility.
+    field_profile = medc.DataArrayInt(restricted_nodes.astype(np.int32))
 
     # Med support mesh for field ( restricted to profile nodes ) without cells
     field_mesh = medc.MEDCouplingUMesh()
@@ -161,7 +163,8 @@ def toMCFieldAndProfileElem(asfield, medmesh, prefix=""):
     medc_cell_field.setArray(field_values)
     medc_cell_field.setNature(medc.IntensiveConservation)
 
-    field_profile = medc.DataArrayInt(restricted_cells)
+    # medcoupling DataArrayInt expects INT32 numpy arrays; cast from int64.
+    field_profile = medc.DataArrayInt(restricted_cells.astype(np.int32))
     field_profile.setName("".join((prefix, "ElemProfile")))
 
     if len(restricted_cells) == medmesh.getNumberOfCells():
